Created a StreamProducer interface and renamed the implementation in signalProcessing to ToyStreamProducer. Created a second implementation called AnotherStreamProducer with the code intended for Kafka.
